1.Net.Core换用dotnetty的buff 2.重写组包拆包代码，没有了额外存储区，也提升了一些性能. 3.修复了字符编码多字节的乱码问题 4.RPC解决了跨平台数据类型不兼容问题 5.每个RPC服务具备独立的数据类型组 6.设计了一层自定义公共数据类型，高效转型，另可根据平台的不同、序列化的差异进行自制Convert。 7.完善了RPC的异常检测与处理 网络框架待完善项:心跳包、TCP异常检测与处理 tip:C# Net.Core和Java Android 已经彻底互通^ ^

+++ Make/MODEL/RPC/Request/ICommand.cs
@@ -7,6 +7,6 @@ namespace Make.MODEL.RPC.Request
 {
     public interface ICommand
     {
-        public void AddHp(Token token, long num);
+        public void AddHp(Token token, int num);
     }
 }
